requirements.txt
streamlit
python-dotenv
langchain
python-docx
PyMuPDF
openai
chromadb
tiktoken
pypdf